requirements.txt
flask==3.1.2
numpy==2.3.4
pydantic==2.12.3
Flask-SQLAlchemy==3.1.1
sqlalchemy==2.0.44
psycopg==3.2.12
python-dotenv==1.2.1
